refactor : 개념에 따라(계산기, 구분자) 패키지화
