+++ src/main/java/calculator/delimiter/Delimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.delimiter;
 
 import java.util.List;
