+++ src/main/java/calculator/delimiter/DelimiterFactory.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.delimiter;
 
 public class DelimiterFactory {
 
@@ -8,17 +8,17 @@ public class DelimiterFactory {
 
     private final DefaultDelimiter defaultDelimiter = new DefaultDelimiter();
 
-    public Delimiter getDelimiter(String input){
+    public Delimiter getDelimiter(String input) {
         // 커스텀 먼저 반드시 확인되어야 한다.
-        if (customDelimiter.applicable(input)){
+        if (customDelimiter.applicable(input)) {
             return customDelimiter;
         }
 
-        if (regularDelimiter.applicable(input)){
+        if (regularDelimiter.applicable(input)) {
             return regularDelimiter;
         }
 
-        if (defaultDelimiter.applicable(input)){
+        if (defaultDelimiter.applicable(input)) {
             return defaultDelimiter;
         }
