+++ src/main/java/calculator/delimiter/RegularDelimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.delimiter;
 
 import java.util.Arrays;
 import java.util.List;
