+++ src/test/java/calculator/CalculatorTest.java
@@ -2,6 +2,7 @@ package calculator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import calculator.calculator.Calculator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -34,7 +35,7 @@ class CalculatorTest {
     }
 
     @ParameterizedTest
-    @CsvSource(delimiter = '=', value =  {"1,  2=3", "      1=1", "   3 ,  1 : 2  =6"})
+    @CsvSource(delimiter = '=', value = {"1,  2=3", "      1=1", "   3 ,  1 : 2  =6"})
     @DisplayName("숫자와 공백이 같이 주어지면 공백을 없애서 계산한다.")
     void replaceEmptyString(String emptyInput, int expected) {
         int calculated = calculator.calculate(emptyInput);
@@ -50,7 +51,8 @@ class CalculatorTest {
 
     @ParameterizedTest
     @DisplayName("쉼표나, 콜론을 포함한 문자열이 주어지면, 쉼포 혹은 콜론을 기준으로 구분하여 숫자로 변환 후 더한다.")
-    @CsvSource(delimiter = '=', value = {"1,2,3=6", "0,1,2=3", "100,200,300=600", "1,2=3", "1:2=3", "1:2,3=6", "100,200:300=600"})
+    @CsvSource(delimiter = '=', value = {"1,2,3=6", "0,1,2=3", "100,200,300=600", "1,2=3", "1:2=3", "1:2,3=6",
+            "100,200:300=600"})
     void calculateByRestOrColon(String input, int expected) {
         int result = calculator.calculate(input);
         Assertions.assertThat(result).isEqualTo(expected);
