+++ src/test/java/calculator/DelimiterFactoryTest.java
@@ -3,6 +3,8 @@ package calculator;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import calculator.delimiter.Delimiter;
+import calculator.delimiter.DelimiterFactory;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
